DeepAI-main/requirements.txt
# Machine Learning Core
lightgbm==4.1.0
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.3
scipy==1.11.4

# Imbalanced Learning
imbalanced-learn==0.11.0

# Explainability
shap==0.43.0

# Reinforcement Learning
torch==2.1.1
gymnasium==0.29.1

# Web Scraping & Network
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
pyOpenSSL==23.3.0
cryptography==41.0.7
python-whois==0.8.0

# Data Validation
pydantic==2.5.2
validators==0.22.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2
uuid==1.30

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Security
bandit==1.7.5
safety==2.3.5

# API (opcional)
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0

# Visualization (notebooks)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
jupyter==1.0.0
ipython==8.19.0
